Alternative Login Platform - modifications
1. Moved link.html to 'users/client'

2. Moved link.js to 'users/client'

3.  link.js: imported Meteor package from 'meteor/meteor'
imported Template from 'meteor/templating'

4. created new file config.js under 'users/server' that contains the logic to set values of github clientId and secret.
The values are read from the collection LoginPlatforms and validated using githubSettingsValid

5. method.js: added function to unlink GitHub credentials from apinf user account.
